rename risk assessment files
